requirements.txt
# Core Dependencies
# =================

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
datasets>=2.14.0          # HuggingFace datasets for WMT14
tokenizers>=0.13.0        # Fast BPE tokenization
sentencepiece>=0.1.99     # Alternative tokenization
sacrebleu>=2.3.0          # BLEU score calculation

# Training Utilities
tensorboard>=2.13.0       # Training visualization
wandb>=0.15.0             # Experiment tracking (optional)
tqdm>=4.65.0              # Progress bars
pyyaml>=6.0               # Config file parsing

# NLP Utilities
nltk>=3.8.0               # NLP utilities
spacy>=3.6.0              # Text processing

# Jupyter & Visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Utilities
requests>=2.31.0
pandas>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Mixed Precision Training
# (already included in torch>=1.6)

# Optional: Distributed Training
# torch-distributed>=0.1.0
